security/auth.py
import jwt
from jwt import PyJWTError
from datetime import datetime,timedelta
import datetime as dt
from fastapi import Depends,HTTPException,status
from fastapi.security import OAuth2PasswordBearer
from typing import Optional
from models import User
from schemas import TokenData
from settings import ENVIRONMENT
from services import UserService,get_user_service

oauth2_scheme = OAuth2PasswordBearer(tokenUrl=f'{ENVIRONMENT.GLOBAL_API_PREFIX}/users/token')

def create_access_token(data:dict,expires_delta:Optional[timedelta] = None) -> str:
    '''
    create a new acces token
    '''
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(dt.timezone.utc) + expires_delta
    else:
        expire = datetime.now(dt.timezone.utc) + timedelta(minutes=ENVIRONMENT.TOKEN_LIFE_TIME)
    to_encode.update({'exp':expire})
    encoded_jwt = jwt.encode(to_encode,ENVIRONMENT.SECRET_KEY,algorithm=ENVIRONMENT.ALGORITHM)
    return encoded_jwt

async def get_current_user(
        token:str = Depends(oauth2_scheme),
        service:UserService = Depends(get_user_service)
) -> User:
    '''
    gets the current user from the authorixation token
    '''
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail='Could not validate credentials',
        headers={'WWW-Authenticate':'Bearer'}
    )
    try:
        payload = jwt.decode(token,ENVIRONMENT.SECRET_KEY,algorithms=[ENVIRONMENT.ALGORITHM])
        username:str = payload.get('sub')
        if username is None:
            raise credentials_exception
        token_data = TokenData(username=username)
    except PyJWTError:
        raise credentials_exception
    
    user = await service.get_by_name(token_data.username) # type: ignore
    if user is None:
        raise credentials_exception
    return user